requirements.txt
# LangChain
langchain>=0.1.14
langchain-community>=0.0.38
langchain-google-genai>=0.0.9
langchain_community>=0.0.38
transformers>=4.39.3

# ChromaDB 
chromadb>=0.4.22
langchain-chroma>=0.1.0


# Sentence Transformers
sentence-transformers>=2.5.1


# Xử lý dữ liệu
pandas>=2.2.1
numpy>=1.26.4

# Google Gemini API
google-generativeai>=0.4.0

# .env
python-dotenv>=1.0.1

# Chạy API nếu muốn mở rộng ứng dụng thành chatbot web
fastapi>=0.110.0
uvicorn>=0.29.0
